Allow slot names for masks: http://bugs.gentoo.org/show_bug.cgi?id=181244
git-svn-id: https://svn.gentooexperimental.org/eix/trunk@415 3c3e6e0f-8f03-0410-a2c1-9066eb21f691

+++ src/portage/basicversion.cc
@@ -120,7 +120,7 @@ BasicVersion::defaults()
 }
 
 void
-BasicVersion::parseVersion(const char *str, int n)
+BasicVersion::parseVersion(const char *str, size_t n)
 {
 	defaults();
 	if(n > 0)
